refactor nautilus prediction engines
