+++ common/src/main/java/ac/grim/grimac/predictionengine/predictions/rideable/PredictionEngineRideableNormal.java
@@ -3,6 +3,7 @@ package ac.grim.grimac.predictionengine.predictions.rideable;
 import ac.grim.grimac.player.GrimPlayer;
 import ac.grim.grimac.predictionengine.predictions.PredictionEngineNormal;
 import ac.grim.grimac.utils.data.VectorData;
+import ac.grim.grimac.utils.math.GrimMath;
 import ac.grim.grimac.utils.math.Vector3dm;
 import lombok.RequiredArgsConstructor;
 
@@ -20,6 +21,18 @@ public class PredictionEngineRideableNormal extends PredictionEngineNormal {
 
     @Override
     public List<VectorData> applyInputsToVelocityPossibilities(GrimPlayer player, Set<VectorData> possibleVectors, float speed) {
-        return PredictionEngineRideableUtils.applyInputsToVelocityPossibilities(movementVector, player, possibleVectors, speed);
+        return PredictionEngineRideableUtils.applyInputsToVelocityPossibilities(this, movementVector, player, possibleVectors, speed);
+    }
+
+    @Override
+    public Vector3dm getMovementResultFromInput(GrimPlayer player, Vector3dm inputVector, float flyingSpeed, float yRot) {
+        float yRotRadians = GrimMath.radians(yRot);
+        float sin = player.trigHandler.sin(yRotRadians);
+        float cos = player.trigHandler.cos(yRotRadians);
+
+        double xResult = inputVector.getX() * cos - inputVector.getZ() * sin;
+        double zResult = inputVector.getZ() * cos + inputVector.getX() * sin;
+
+        return new Vector3dm(xResult * flyingSpeed, inputVector.getY() * flyingSpeed, zResult * flyingSpeed);
     }
 }
