+++ common/src/main/java/ac/grim/grimac/predictionengine/predictions/rideable/PredictionEngineRideableWater.java
@@ -3,6 +3,7 @@ package ac.grim.grimac.predictionengine.predictions.rideable;
 import ac.grim.grimac.player.GrimPlayer;
 import ac.grim.grimac.predictionengine.predictions.PredictionEngineWater;
 import ac.grim.grimac.utils.data.VectorData;
+import ac.grim.grimac.utils.math.GrimMath;
 import ac.grim.grimac.utils.math.Vector3dm;
 import lombok.RequiredArgsConstructor;
 
@@ -11,7 +12,7 @@ import java.util.Set;
 
 @RequiredArgsConstructor
 public class PredictionEngineRideableWater extends PredictionEngineWater {
-    private final Vector3dm movementVector;
+    protected final Vector3dm movementVector;
 
     @Override
     public void addJumpsToPossibilities(GrimPlayer player, Set<VectorData> existingVelocities) {
@@ -20,6 +21,19 @@ public class PredictionEngineRideableWater extends PredictionEngineWater {
 
     @Override
     public List<VectorData> applyInputsToVelocityPossibilities(GrimPlayer player, Set<VectorData> possibleVectors, float speed) {
-        return PredictionEngineRideableUtils.applyInputsToVelocityPossibilities(movementVector, player, possibleVectors, speed);
+        return PredictionEngineRideableUtils.applyInputsToVelocityPossibilities(this, movementVector, player, possibleVectors, speed);
     }
+
+    @Override
+    public Vector3dm getMovementResultFromInput(GrimPlayer player, Vector3dm inputVector, float flyingSpeed, float yRot) {
+        float yRotRadians = GrimMath.radians(yRot);
+        float sin = player.trigHandler.sin(yRotRadians);
+        float cos = player.trigHandler.cos(yRotRadians);
+
+        double xResult = inputVector.getX() * cos - inputVector.getZ() * sin;
+        double zResult = inputVector.getZ() * cos + inputVector.getX() * sin;
+
+        return new Vector3dm(xResult * flyingSpeed, inputVector.getY() * flyingSpeed, zResult * flyingSpeed);
+    }
+
 }
